refactor: split deploy between remote and local
