+++ cmd/deploy/remote_deploy.go
@@ -1,0 +1,29 @@
+// Copyright 2023 The Okteto Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package deploy
+
+import "context"
+
+type remoteDeployCommand struct {
+}
+
+func newRemoteDeployer() *remoteDeployCommand {
+	return &remoteDeployCommand{}
+}
+
+func (rd *remoteDeployCommand) deploy(ctx context.Context, deployOptions *Options) error {
+	return nil
+}
+
+func (rd *remoteDeployCommand) cleanUp(ctx context.Context, err error) {}
